requirements.txt
flask==3.0.3
python-dotenv==1.0.1
azure-identity==1.17.1
azure-storage-blob==12.22.0
azure-storage-queue==12.11.0
PyPDF2==3.0.1
pdf2image==1.17.0
Pillow==10.4.0
uuid==1.30
azure-ai-documentintelligence==1.0.0b3
flask-cors==4.0.1
pyautogen==0.2.33
pytest==8.3.2
opentelemetry-instrumentation-flask==0.47b0
langchain==0.2.12
langchain-openai==0.1.20
tenacity==8.5.0
flask-socketio==5.3.6
gunicorn==22.0.0
gevent==24.2.1
graphrag==0.2.0
pandas==2.2.2
adlfs==2024.7.0
fsspec==2024.6.1
azure-data-tables==12.5.0
flask[async]==3.0.3
gevent-websocket==0.10.1
asgiref==3.8.1
nest_asyncio==1.6.0
future==1.0.0
graspologic==3.4.1
hyppo==0.4.0
instructor
faiss-cpu==1.8.0.post1
sentence-transformers
huggingface_hub
